% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_graph.R
\name{correlation_graph}
\alias{correlation_graph}
\title{Create a ggpair scatter plot}
\usage{
correlation_graph(dataframe)
}
\arguments{
\item{dataframe}{A data frame or tibble.}
}
\value{
A ggpairs scatterplot representing a correlation matrix between pairs of variables
}
\description{
A short-cut function for creating 2 dimensional ggpair scatterplots
}
\examples{
correlation_graph(mtcars)
}
