% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{data_cleaning}
\alias{data_cleaning}
\title{Creates a new data frame with selected columns dropped from the old data frame}
\usage{
data_cleaning(data_frame, columns)
}
\arguments{
\item{data_frame}{A data frame or data frame extension (e.g. a tibble).}

\item{columns}{Name(s) of column(s) from the data frame}
}
\value{
A new data frame with columns dropped from the old data frame.
If all columns are to be removed, an empty data frame will be returned.
}
\description{
Creates a new data frame with selected columns dropped from the old data frame
}
\examples{
data_cleaning(mtcars, c("mpg","disp","qsec"))
data_cleaning(mtcars, "disp")
}
