% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe_scale_center.R
\name{recipe_scale_center}
\alias{recipe_scale_center}
\title{Create a recipe to preprocess data}
\usage{
recipe_scale_center(dataframe, formula)
}
\arguments{
\item{dataframe}{A data frame or a tibble of the data.}

\item{formula}{The formula used to create the model (eg: target ~ predictor).}
}
\value{
A tidymodel recipe, specifying inputs, outputs, and steps
(scaling and centering) applied to all predictors.
}
\description{
Creates a recipe by scaling and centering the given data,
and creates a model using the given formula.
}
\examples{
recipe_scale_center(mtcars, mpg ~ hp)
recipe_scale_center(mtcars, mpg ~ hp + cyl)
recipe_scale_center(mtcars, mpg ~ .)
}
